Execute all datastore db operations on a single thread
All database operations are performed on a single operation thread,
this stops leaking of connections when a user's application uses the
datastore on two different threads, but only calls close on a single
thread.

Remove threadlocal code from SQLiteWrapper, enforcing the need to perform
each database operation on the same thread.

Added a TestUtil method to create a connection to underlying database
for the datastore.

Remove test which checks if the underlying db is closed when
datastore.close() is called, this test can't be implemented since the
database can only be called on the executor service thread.

Ensure Query and indexing db accesses all happen on executor thread

Add functions to DatastoreExtended to control transaction state